refactor: share day trade pnl calc
