+++ apps/web/app/lib/metrics.ts
@@ -1,6 +1,8 @@
 import type { EnrichedTrade } from "@/lib/fifo";
 import type { Position } from '@/lib/services/dataService';
 import { nowNY } from '@/lib/timezone';
+import { calcTodayTradePnL } from "./calcTodayTradePnL";
+export { calcTodayTradePnL } from "./calcTodayTradePnL";
 
 // Only enable verbose logging outside production
 const DEBUG = process.env.NODE_ENV !== 'production';
@@ -129,69 +131,7 @@ function sum(arr: number[]): number {
  * @param todayStr 今日日期字符串，格式为 YYYY-MM-DD
  * @returns 日内交易盈亏
  */
-function calcTodayTradePnL(enrichedTrades: EnrichedTrade[], todayStr: string): number {
-  // 为多头和空头分别维护栈
-  const longMap: Record<string, { qty: number; price: number }[]> = {};
-  const shortMap: Record<string, { qty: number; price: number }[]> = {};
-  let pnl = 0;
-
-  // 按时间顺序处理今日交易
-  enrichedTrades
-    .filter(t => t.date.startsWith(todayStr))
-    .sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime())
-    .forEach(t => {
-      const { symbol, action, quantity, price } = t;
-
-      // 初始化栈
-      if (!longMap[symbol]) longMap[symbol] = [];
-      if (!shortMap[symbol]) shortMap[symbol] = [];
-
-      if (action === 'buy') {
-        // 买入：直接加入多头栈
-        longMap[symbol].push({ qty: quantity, price });
-      }
-      else if (action === 'sell') {
-        // 卖出：匹配今日多头栈
-        const longStack = longMap[symbol];
-        let remain = quantity;
-
-        while (remain > 0 && longStack.length > 0) {
-          const batch = longStack[0]!;
-          const q = Math.min(batch.qty, remain);
-          // 多头平仓：卖出价 > 买入价 = 盈利
-          pnl += (price - batch.price) * q;
-          batch.qty -= q;
-          remain -= q;
-          if (batch.qty === 0) longStack.shift();
-        }
-
-        // 剩余的不处理（可能是平历史仓位，不计入日内交易）
-      }
-      else if (action === 'short') {
-        // 做空：直接加入空头栈
-        shortMap[symbol].push({ qty: quantity, price });
-      }
-      else if (action === 'cover') {
-        // 回补：匹配今日空头栈
-        const shortStack = shortMap[symbol];
-        let remain = quantity;
-
-        while (remain > 0 && shortStack.length > 0) {
-          const batch = shortStack[0]!;
-          const q = Math.min(batch.qty, remain);
-          // 空头平仓：回补价 < 做空价 = 盈利
-          pnl += (batch.price - price) * q;
-          batch.qty -= q;
-          remain -= q;
-          if (batch.qty === 0) shortStack.shift();
-        }
-
-        // 剩余的不处理（可能是平历史仓位，不计入日内交易）
-      }
-    });
-
-  return pnl;
-}
+// calcTodayTradePnL moved to its own module for reuse
 
 /**
  * 计算日内交易盈亏（FIFO视角）
